requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
discord.py==2.5.2
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
feedparser==6.0.11
fiona==1.10.1
fonttools==4.59.0
frozenlist==1.7.0
geopandas==1.1.1
icalendar==6.3.1
ics==0.7.2
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.1.3
multidict==6.6.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyogrio==0.11.1
pyparsing==3.2.3
pyproj==3.7.1
python-dotenv
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
recurring_ical_events==3.4.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
rtree==1.4.0
sgmllib3k==1.0.0
shapely==2.1.1
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
TatSu==5.13.1
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
x_wr_timezone==2.0.1
yarg==0.1.9
yarl==1.20.1
asyncpg==0.31.0

# Web Interface
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.27.0
